chore: expose errors
